// src/interfaces/entry-transfers.interface.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface EntryTransfers {
  wildcards?: {
    playedTimeFormatted?: string;
    status?: string;
    name?: string;
    time?: string;
    chip?: number;
    entry?: number;
    event?: number;
    [k: string]: any;
  }[];
  entry?: {
    id?: number;
    playerFirstName?: string;
    playerLastName?: string;
    playerRegionId?: number;
    playerRegionName?: string;
    playerRegionShortIso?: string;
    summaryOverallPoints?: number;
    summaryOverallRank?: number;
    summaryEventPoints?: number;
    summaryEventRank?: number;
    joinedSeconds?: number;
    currentEvent?: number;
    totalTransfers?: number;
    totalLoans?: number;
    totalLoansActive?: number;
    transfersOrLoans?: string;
    deleted?: boolean;
    email?: boolean;
    joinedTime?: string;
    name?: string;
    bank?: number;
    value?: number;
    kit?: null;
    eventTransfers?: number;
    eventTransfersCost?: number;
    extraFreeTransfers?: number;
    strategy?: null;
    favouriteTeam?: null;
    startedEvent?: number;
    player?: number;
    [k: string]: any;
  };
  leagues?: {
    cup?: any[];
    h2h?: any[];
    classic?: {
      id?: number;
      entryRank?: number;
      entryLastRank?: number;
      entryMovement?: string;
      entryChange?: number | null;
      entryCanLeave?: boolean;
      entryCanAdmin?: boolean;
      entryCanInvite?: boolean;
      entryCanForum?: boolean;
      entryCode?: null | string;
      name?: string;
      shortName?: string | null;
      created?: string;
      closed?: boolean;
      forumDisabled?: boolean;
      makeCodePublic?: boolean;
      rank?: null;
      size?: null;
      leagueType?: string;
      scoring?: string;
      reprocessStandings?: boolean;
      adminEntry?: null | number;
      startEvent?: number;
      [k: string]: any;
    }[];
    [k: string]: any;
  };
  history?: {
    id?: number;
    timeFormatted?: string;
    time?: string;
    elementInCost?: number;
    elementOutCost?: number;
    elementIn?: number;
    elementOut?: number;
    entry?: number;
    event?: number;
    [k: string]: any;
  }[];
  [k: string]: any;
}